+++ src/main/java/conspire/powers/ExplosiveBarrierPower.java
@@ -8,21 +8,17 @@ import com.megacrit.cardcrawl.core.AbstractCreature;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.localization.PowerStrings;
-import com.megacrit.cardcrawl.powers.AbstractPower;
 
-public class ExplosiveBarrierPower extends AbstractPower {
+public class ExplosiveBarrierPower extends AbstractConspirePower {
     public static final String POWER_ID = "conspire:ExplosiveBarrier";
     private static final PowerStrings powerStrings = CardCrawlGame.languagePack.getPowerStrings(POWER_ID);
     public static final String NAME = powerStrings.NAME;
     public static final String[] DESCRIPTIONS = powerStrings.DESCRIPTIONS;
 
     public ExplosiveBarrierPower(AbstractCreature owner, int amount) {
-        this.name = NAME;
-        this.ID = POWER_ID;
-        this.owner = owner;
+        super(POWER_ID, NAME, owner);
         this.amount = amount;
         this.updateDescription();
-        this.loadRegion("explosive");
     }
 
     @Override